requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.40.0
pydantic>=2.8.0
pytest>=8.3.0
langgraph>=0.2.50
clickhouse-connect>=0.8.15
groq>=0.13.1
pandas>=2.2.3
python-dotenv>=1.0.1
chromadb>=0.5.23
pypdf>=5.1.0
python-docx>=1.1.2
